Some more WIP improvements for the Axum example
